app/src/main/java/com/example/w/startkotlin/domain/datasource/ForecastProvider.kt
package com.example.w.startkotlin.domain.datasource

import com.example.w.startkotlin.data.db.ForecastDb
import com.example.w.startkotlin.data.server.ForecastServer
import com.example.w.startkotlin.domain.model.Forecast
import com.example.w.startkotlin.domain.model.ForecastList
import com.example.w.startkotlin.extensions.firstResult

class ForecastProvider(private val sources: List<ForecastDataSource> = ForecastProvider.SOURCES) {

    companion object {
        val DAY_IN_MILLIS = 1000 * 60 * 60 * 24
        val SOURCES by lazy { listOf(ForecastDb(), ForecastServer()) }
    }

    fun requestByZipCode(zipCode: Long, days: Int): ForecastList = requestToSources {
        val res = it.requestForecastByZipCode(zipCode, todayTimeSpan())
        if (res != null && res.size >= days) res else null
    }

    fun requestForecast(id: Long): Forecast = requestToSources { it.requestDayForecast(id) }

    private fun todayTimeSpan() = System.currentTimeMillis() / DAY_IN_MILLIS * DAY_IN_MILLIS

    private fun <T : Any> requestToSources(f: (ForecastDataSource) -> T?): T = sources.firstResult { f(it) }

}